+++ lib/commandline/cexport/cexport.macos.ml
@@ -21,13 +21,6 @@ let name = CexportCommon.name
 
 type t = CexportCommon.export_t
 
-let term_website = CexportCommon.term_website
-let term_regex = CexportCommon.term_regex
-let term_paste = CexportCommon.term_paste
-let term_output = CexportCommon.term_output
-let term_display_time = CexportCommon.term_display_time
-let term_show_password = CexportCommon.term_show_password
-
 let validate t =
   let () = Libcithare.Manager.check_initialized () in
   let () =
@@ -41,7 +34,7 @@ let validate t =
   in
   ()
 
-let fpaste ~regex ~paste password =
+let fpaste ?mail ?username ~regex ~paste password =
   match paste with
   | false ->
       let () = print_endline password.Libcithare.Password.password in
@@ -59,6 +52,17 @@ let fpaste ~regex ~paste password =
                 ()
             in
             let () =
+              if Option.is_some mail then
+                Printf.printf "For : %s\n"
+                @@ Option.value ~default:String.empty password.mail
+            in
+            let () =
+              if Option.is_some username then
+                Printf.printf "For : %s\n"
+                @@ Option.value ~default:String.empty password.username
+            in
+
+            let () =
               Printf.printf "Password successfully written in pasteboard\n"
             in
             ()
@@ -66,20 +70,18 @@ let fpaste ~regex ~paste password =
 
       ()
 
-let term_cmd () =
-  let combine website regex paste output =
-    let export =
-      new CexportCommon.export_t validate fpaste website regex paste output
-    in
-    export#run ()
-  in
-  Term.(const combine $ term_website $ term_regex $ term_paste $ term_output)
+let term_paste =
+  Arg.(
+    value & flag
+    & info [ "p"; "paste" ] ~doc:"Write the password into the pasteboard"
+  )
 
+let term_cmd = CexportCommon.term_cmd ~term_paste validate fpaste
 let doc = CexportCommon.doc
-let man = [ `S Manpage.s_description; `P "Export passwords" ]
+let man = CexportCommon.man
 
 let cmd () =
   let info = Cmd.info ~doc ~man name in
-  Cmd.v info (term_cmd ())
+  Cmd.v info term_cmd
 
 let command = cmd ()