+++ lib/util/util.ml
@@ -18,3 +18,4 @@
 module FileSys = FileSys
 module Io = Io
 module Ustring = Ustring
+module Misc = Misc